revert(refactor): rewrite in rust
This reverts commit d7d7c37ff18ff7f1bd1b947e5bfc6c55867eb689.

+++ bench/typings/stringzilla.pyi
@@ -1,5 +1,3 @@
 from collections.abc import Buffer
 
-from cityhash import Uint64
-
-def hash(data: str | Buffer, /, seed: int = ...) -> Uint64: ...  # noqa: A001
+def hash(data: str | Buffer, /, seed: int = ...) -> int: ...  # noqa: A001
